+++ CHANGELOG.md
@@ -56,6 +56,8 @@
   depending on the length of the content, see [#359](https://github.com/cadeyrn/enterprise-policy-generator/issues/359)
 - The button for selecting the generated code now copies it directly to the clipboard, see
   [#364](https://github.com/cadeyrn/enterprise-policy-generator/issues/364)
+- When attempting to load or import an incompatible configuration, a corresponding warning is displayed, see
+  [#362](https://github.com/cadeyrn/enterprise-policy-generator/issues/362)
 - Added a promo box to give users of the extension an easy way to support the development of this extension, see
   [#360](https://github.com/cadeyrn/enterprise-policy-generator/issues/360)